04-data-engineering-mlops/15-data-pipeline/requirements.txt
requests==2.26.0
pandas==1.3.4
numpy==1.21.4
datetime==4.3
sqlalchemy==1.4.31
psycopg2-binary==2.9.1
apache-airflow==2.2.3
email-validator==1.1.3
markupsafe==1.1.1
